1.加强用户密码安全，使用bcryptjs算法

+++ models/posts.js
@@ -1,5 +1,6 @@
-const marked = require('marked')
-const Post = require('../lib/mongo').Post
+const marked = require('marked');
+const Post = require('../lib/mongo').Post;
+const CommentModel = require('./comments');
 
 // 将 post 的 content 从 markdown 转换成 html
 Post.plugin('contentToHtml', {
@@ -15,7 +16,31 @@ Post.plugin('contentToHtml', {
         }
         return post
     }
-})
+});
+
+// 给 post 添加留言数 commentsCount
+Post.plugin('addCommentsCount', {
+    afterFind : function (posts) {
+        return Promise.all(posts.map(function (post) {
+            return CommentModel.getCommentsCount(post._id)
+                .then(function (commentsCount) {
+                    post.commentsCount = commentsCount;
+                    return post;
+                })
+        }))
+    },
+
+    afterFindOne : function (post) {
+        if(post){
+            return CommentModel.getCommentsCount(post._id)
+                .then(function (commentsCount) {
+                    post.commentsCount = commentsCount;
+                    return post;
+                })
+        }
+        return post;
+    },
+});
 
 module.exports = {
     // 创建一篇文章
@@ -29,6 +54,7 @@ module.exports = {
             .findOne({ _id: postId })
             .populate({ path: 'author', model: 'User' })
             .addCreatedAt()
+            .addCommentsCount()
             .contentToHtml()
             .exec()
     },
@@ -44,6 +70,7 @@ module.exports = {
             .populate({ path: 'author', model: 'User' })
             .sort({ _id: -1 })
             .addCreatedAt()
+            .addCommentsCount()
             .contentToHtml()
             .exec()
     },
@@ -70,10 +97,16 @@ module.exports = {
             .exec()
     },
 
-    // 通过文章 id 删除一篇文章
-    delPostById: function delPostById (postId) {
+    // 通过用户 id 和文章 id 删除一篇文章
+    delPostById: function delPostById (postId, author) {
         return Post
-            .deleteOne({ _id: postId })
+            .deleteOne({ _id: postId, author : author })
             .exec()
+            .then(function (value) {
+                // 文章删除后，再删除该文章下的所有留言
+                if (res.result.ok && res.result.n > 0) {
+                    return CommentModel.delCommentsByPostId(postId);
+                }
+            })
     }
 }
